+++ ingress_routes.tf
@@ -1,16 +1,31 @@
+/**
+ * Copyright (C) 2023 Solution Libre <contact@solution-libre.fr>
+ * 
+ * This file is part of Traefik Terraform module.
+ * 
+ * Traefik Terraform module is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ * 
+ * Traefik Terraform module is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ * 
+ * You should have received a copy of the GNU General Public License
+ * along with Traefik Terraform module.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
 resource "kubernetes_manifest" "ingress_routes" {
   for_each = var.ingress_routes
 
   manifest = yamldecode(templatefile(
     "${path.module}/templates/manifests/ingress-route.yaml.tpl",
-    {
-      name        = each.key
-      namespace   = each.value.namespace
-      middlewares = each.value.middlewares
-      hostname    = each.value.hostname
-      service     = each.value.service
-      redirect    = each.value.redirect
-    }
+    merge(
+      { name = each.key },
+      { for k, v in each.value : k => v }
+    )
   ))
 
   depends_on = [