+++ outputs.tf
@@ -1,3 +1,22 @@
+/**
+ * Copyright (C) 2023 Solution Libre <contact@solution-libre.fr>
+ * 
+ * This file is part of Traefik Terraform module.
+ * 
+ * Traefik Terraform module is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ * 
+ * Traefik Terraform module is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ * 
+ * You should have received a copy of the GNU General Public License
+ * along with Traefik Terraform module.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
 output "namespace" {
   description = "Traefik namespace"
   value       = module.generic.namespace
@@ -5,5 +24,5 @@ output "namespace" {
 
 output "version" {
   description = "Traefik version"
-  value       = module.generic.app_version
+  value       = module.generic.version
 }